feat(zink): expand module storage
