+++ zink/src/storage/mod.rs
@@ -1,0 +1,7 @@
+//! Zink storage implementation.
+
+pub use {array::StorageArray, kv::Storage, mapping::StorageMapping};
+
+mod array;
+mod kv;
+mod mapping;
